# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Docuflow API
  version: 0.2.0
  description: Local-first Go API. Original PDFs are immutable and processing outputs are versions.
servers:
  - url: /api
paths:
...
  /documents:
    get:
      operationId: listDocuments
      responses:
        "200":
          description: Document dashboard data
          content:
            application/json:
              schema:
                type: object
                required: [documents]
                properties:
                  documents: { type: array, items: { $ref: "#/components/schemas/Document" } }
        default: { $ref: "#/components/responses/Error" }
    post:
      operationId: uploadDocument
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        "201":
          description: Immutable original and upload version created
          content:
            application/json:
              schema:
                type: object
                required: [document]
                properties:
                  document: { $ref: "#/components/schemas/Document" }
        default: { $ref: "#/components/responses/Error" }
  /edit-sessions:
    post:
      operationId: createEditSession
      description: Upload one PDF directly and create a Preview session. The session ID is an internal document ID; no prior Recent Files selection is required.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        "201":
          description: Preview session created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EditSessionEnvelope" }
        default: { $ref: "#/components/responses/Error" }
  /edit-sessions/{sessionId}:
    parameters:
      - name: sessionId
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      operationId: getEditSession
      responses:
        "200":
          description: Existing Preview session
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EditSessionEnvelope" }
        default: { $ref: "#/components/responses/Error" }
  /edit-sessions/{sessionId}/export:
    parameters:
      - name: sessionId
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      operationId: exportEditSession
      description: >-
        Flattens an overlay editor document onto the session PDF and stores the
        result as a new version. The original is never modified. Send JSON when
        the document has no images, or multipart with the JSON in a "document"
        field and each JPEG in a part named after the asset it fulfils.
        Native content editing remains unavailable and is reported separately by
        capabilities.nativeContentEditing.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AnnotationDocument" }
          multipart/form-data:
            schema:
              type: object
              required: [document]
              properties:
                document:
                  description: The AnnotationDocument encoded as a JSON string.
                  type: string
              additionalProperties: { type: string, format: binary }
      responses:
        "201":
          description: The annotated version was saved
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DirectToolResult" }
        default: { $ref: "#/components/responses/Error" }
  /documents/{documentId}/fonts:
    parameters: [{ $ref: "#/components/parameters/DocumentId" }]
    get:
      operationId: getDocumentFonts
      description: >-
        Lists the typefaces the PDF already carries, as reported by pdffonts.
        The editor pairs these against the registry so a font the server cannot
        reproduce can be named precisely.
      responses:
        "200":
          description: Fonts embedded in or referenced by the document
          content:
            application/json:
              schema:
                type: object
                required: [fonts]
                properties:
                  fonts:
                    type: array
                    items:
                      type: object
                      required: [name, type, embedded, subset]
                      properties:
                        name: { type: string }
                        type: { type: string, description: PDF font type, for example "CID TrueType" }
                        embedded: { type: boolean }
                        subset: { type: boolean }
        default: { $ref: "#/components/responses/Error" }
  /fonts:
    get:
      operationId: listFonts
      description: >-
        Lists the embeddable TrueType fonts the server found in FONT_DIR, plus
        the files it refused and why. An empty list means only the built-in
        Helvetica is available, which covers Latin-1 only.
      responses:
        "200":
          description: Registered fonts and rejected files
          content:
            application/json:
              schema:
                type: object
                required: [fonts, issues]
                properties:
                  fonts: { type: array, items: { $ref: "#/components/schemas/RegisteredFont" } }
                  issues:
                    type: array
                    items:
                      type: object
                      required: [file, reason]
                      properties:
                        file: { type: string }
                        reason: { type: string }
        default: { $ref: "#/components/responses/Error" }
  /documents/trash:
    get:
      operationId: listTrash
      description: Lists soft-deleted documents whose immutable bytes are still retained.
      responses:
        "200":
          description: Recoverable Trash records
          content:
            application/json:
              schema:
                type: object
                required: [documents]
                properties:
                  documents: { type: array, items: { $ref: "#/components/schemas/Document" } }
        default: { $ref: "#/components/responses/Error" }
  /documents/bulk-delete:
    post:
      operationId: bulkDeleteDocuments
      description: >-
        Moves several documents to Trash in one call. Originals and versions stay
        on disk. The response is itemised: a document that is already gone is
        reported as a failure without affecting the rest of the selection.
      requestBody: { $ref: "#/components/requestBodies/BulkDocuments" }
      responses:
        "200": { $ref: "#/components/responses/BulkDeleteResult" }
        default: { $ref: "#/components/responses/Error" }
  /documents/bulk-permanent-delete:
    post:
      operationId: bulkPermanentlyDeleteDocuments
      description: >-
        Permanently removes several documents already in Trash, including their
        original and every version. Each document is staged and committed on its
        own, so the ones reported as deleted are fully gone and the ones reported
        as failed keep both their rows and their bytes. This cannot be undone.
      requestBody: { $ref: "#/components/requestBodies/BulkDocuments" }
      responses:
        "200": { $ref: "#/components/responses/BulkDeleteResult" }
        default: { $ref: "#/components/responses/Error" }
  /documents/{documentId}:
    parameters: [{ $ref: "#/components/parameters/DocumentId" }]
    get:
      operationId: getDocument
      responses:
        "200":
          description: Document metadata
          content:
            application/json:
              schema:
                type: object
                required: [document]
                properties:
                  document: { $ref: "#/components/schemas/Document" }
        default: { $ref: "#/components/responses/Error" }
    patch:
      operationId: renameDocument
      description: Changes application display metadata only; the UUID-backed storage path is unchanged.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties: { name: { type: string, maxLength: 240, example: "proposal-final.pdf" } }
      responses:
        "200": { description: Display name updated }
        default: { $ref: "#/components/responses/Error" }
    delete:
      operationId: deleteDocument
      description: Soft deletes document metadata from Recent Files. Immutable original and version bytes are retained.
      responses:
        "204": { description: Document removed from Recent Files }
        default: { $ref: "#/components/responses/Error" }
  /documents/{documentId}/restore:
    parameters: [{ $ref: "#/components/parameters/DocumentId" }]
    post:
      operationId: restoreDocument
      description: Restores a soft-deleted document to Recent Files.
      responses:
        "204": { description: Document restored }
        default: { $ref: "#/components/responses/Error" }
  /documents/{documentId}/permanent:
    parameters: [{ $ref: "#/components/parameters/DocumentId" }]
    delete:
      operationId: permanentlyDeleteDocument
      description: Permanently removes a Trash record, its original PDF, and every stored version. Active documents cannot use this endpoint.
      responses:
        "204": { description: Document and stored bytes permanently removed }
        default: { $ref: "#/components/responses/Error" }
...
  /documents/{documentId}/versions:
    parameters: [{ $ref: "#/components/parameters/DocumentId" }]
    get:
      operationId: listDocumentVersions
      responses:
        "200":
          description: Newest-first immutable version history
          content:
            application/json:
              schema:
                type: object
                required: [versions]
                properties:
                  versions: { type: array, items: { $ref: "#/components/schemas/DocumentVersion" } }
        default: { $ref: "#/components/responses/Error" }
  /documents/{documentId}/metadata:
    parameters: [{ $ref: "#/components/parameters/DocumentId" }]
    get:
      operationId: getDocumentMetadata
      responses:
        "200": { description: Editable PDF metadata and read-only document information }
        default: { $ref: "#/components/responses/Error" }
    patch:
      operationId: updateDocumentMetadata
      parameters:
        - name: X-Confirm-Signature-Invalidation
          in: header
          schema: { type: string, enum: ["true", "false"] }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PDFMetadata" }
      responses: { "201": { $ref: "#/components/responses/ProcessingResult" }, default: { $ref: "#/components/responses/Error" } }
  /documents/{documentId}/pages/{page}/thumbnail:
    parameters:
      - { $ref: "#/components/parameters/DocumentId" }
      - { name: page, in: path, required: true, schema: { type: integer, minimum: 1 } }
    get:
      operationId: getPageThumbnail
      responses:
        "200": { description: Lazily rendered PNG thumbnail, content: { image/png: { schema: { type: string, format: binary } } } }
        default: { $ref: "#/components/responses/Error" }
...
  /tools/merge:
    post:
      operationId: mergeDocuments
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [documentIds]
              properties:
                documentIds: { type: array, minItems: 2, maxItems: 20, items: { type: string, format: uuid } }
          multipart/form-data:
            schema:
              type: object
              required: [files]
              properties:
                files:
                  type: array
                  minItems: 2
                  maxItems: 20
                  items: { type: string, format: binary }
      responses: { "201": { $ref: "#/components/responses/ProcessingResult" }, default: { $ref: "#/components/responses/Error" } }
  /tools/split:
    post:
      operationId: splitDocument
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - { $ref: "#/components/schemas/SingleDocumentOperation" }
                - type: object
                  required: [ranges]
                  properties: { ranges: { type: string, example: "1-3,5" } }
          multipart/form-data:
            schema:
              type: object
              required: [file, pages]
              properties:
                file: { type: string, format: binary }
                pages:
                  type: string
                  description: JSON integer array. Each selected page produces one independent one-page PDF output. The legacy plain `ranges` field remains accepted for backwards compatibility.
                  example: '[1,2,3]'
      responses: { "201": { $ref: "#/components/responses/SplitProcessingResult" }, default: { $ref: "#/components/responses/Error" } }
  /tools/extract:
    post:
      operationId: extractPages
      requestBody: { $ref: "#/components/requestBodies/PageRange" }
      responses: { "201": { $ref: "#/components/responses/ProcessingResult" }, default: { $ref: "#/components/responses/Error" } }
  /tools/rotate:
    post:
      operationId: rotatePages
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - { $ref: "#/components/schemas/SingleDocumentOperation" }
                - type: object
                  required: [pages, degrees]
                  properties:
                    pages: { type: array, items: { type: integer, minimum: 1 } }
                    degrees: { type: integer, enum: [-270, -180, -90, 90, 180, 270] }
      responses: { "201": { $ref: "#/components/responses/ProcessingResult" }, default: { $ref: "#/components/responses/Error" } }
  /tools/reorder:
    post:
      operationId: reorderPages
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - { $ref: "#/components/schemas/SingleDocumentOperation" }
                - type: object
                  required: [pageOrder]
                  properties:
                    pageOrder: { type: array, items: { type: integer, minimum: 1 } }
      responses: { "201": { $ref: "#/components/responses/ProcessingResult" }, default: { $ref: "#/components/responses/Error" } }
  /tools/delete-pages:
    post:
      operationId: deletePages
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - { $ref: "#/components/schemas/SingleDocumentOperation" }
                - type: object
                  required: [pages]
                  properties:
                    pages: { type: array, items: { type: integer, minimum: 1 } }
      responses: { "201": { $ref: "#/components/responses/ProcessingResult" }, default: { $ref: "#/components/responses/Error" } }
  /tools/duplicate-pages:
    post:
      operationId: duplicatePages
      requestBody: { $ref: "#/components/requestBodies/PageList" }
      responses: { "201": { $ref: "#/components/responses/ProcessingResult" }, default: { $ref: "#/components/responses/Error" } }
  /tools/insert-pages:
    post:
      operationId: insertPages
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - { $ref: "#/components/schemas/SingleDocumentOperation" }
                - type: object
                  required: [sourceDocumentId, position]
                  properties:
                    sourceDocumentId: { type: string, format: uuid }
                    position: { type: string, enum: [beginning, end, before, after] }
                    page: { type: integer, minimum: 1 }
      responses: { "201": { $ref: "#/components/responses/ProcessingResult" }, default: { $ref: "#/components/responses/Error" } }
  /tools/insert-blank-page:
    post:
      operationId: insertBlankPage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - { $ref: "#/components/schemas/SingleDocumentOperation" }
                - type: object
                  required: [position, pageSize]
                  properties:
                    position: { type: string, enum: [beginning, end, before, after] }
                    page: { type: integer, minimum: 1 }
                    pageSize: { type: string, enum: [same, a4, letter] }
                    orientation: { type: string, enum: [portrait, landscape] }
      responses: { "201": { $ref: "#/components/responses/ProcessingResult" }, default: { $ref: "#/components/responses/Error" } }
  /tools/protect:
    post:
      operationId: protectDocument
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - { $ref: "#/components/schemas/SingleDocumentOperation" }
                - type: object
                  required: [password, confirmPassword]
                  properties:
                    password: { type: string, format: password, writeOnly: true }
                    confirmPassword: { type: string, format: password, writeOnly: true }
                    printing: { type: boolean }
                    copying: { type: boolean }
                    modification: { type: boolean }
                    annotation: { type: boolean }
                    formFilling: { type: boolean }
                    assembly: { type: boolean }
      responses: { "201": { $ref: "#/components/responses/ProcessingResult" }, default: { $ref: "#/components/responses/Error" } }
  /tools/unlock:
    post:
      operationId: unlockDocument
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - { $ref: "#/components/schemas/SingleDocumentOperation" }
                - type: object
                  required: [password]
                  properties: { password: { type: string, format: password, writeOnly: true } }
      responses: { "201": { $ref: "#/components/responses/ProcessingResult" }, default: { $ref: "#/components/responses/Error" } }
  /tools/watermark:
    post:
      operationId: watermarkDocument
      description: JSON applies a text watermark; multipart form data applies a validated JPEG image watermark.
      responses: { "201": { $ref: "#/components/responses/ProcessingResult" }, default: { $ref: "#/components/responses/Error" } }
  /tools/page-numbers:
    post:
      operationId: addPageNumbers
      requestBody: { $ref: "#/components/requestBodies/SingleDocument" }
      responses: { "201": { $ref: "#/components/responses/ProcessingResult" }, default: { $ref: "#/components/responses/Error" } }
  /tools/header-footer:
    post:
      operationId: addHeaderFooter
      requestBody: { $ref: "#/components/requestBodies/SingleDocument" }
      responses: { "201": { $ref: "#/components/responses/ProcessingResult" }, default: { $ref: "#/components/responses/Error" } }
  /tools/compress:
    post:
      operationId: compressDocument
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SingleDocumentOperation" }
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
                mode: { type: string, const: lossless-structural, default: lossless-structural }
      responses: { "201": { $ref: "#/components/responses/ProcessingResult" }, default: { $ref: "#/components/responses/Error" } }
  /tools/ocr:
    post:
      operationId: ocrDocument
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - { $ref: "#/components/schemas/SingleDocumentOperation" }
                - type: object
                  required: [language]
                  properties:
                    language: { type: string, enum: [eng, ind] }
          multipart/form-data:
            schema:
              type: object
              required: [file, language]
              properties:
                file: { type: string, format: binary }
                language: { type: string, enum: [eng, ind], default: eng }
      responses: { "201": { $ref: "#/components/responses/ProcessingResult" }, default: { $ref: "#/components/responses/Error" } }
  /tools/convert/jpg-to-pdf:
    post:
      operationId: convertJpgToPdf
      description: Streams one to 20 validated JPG/JPEG images and creates one immutable PDF. Multipart order becomes PDF page order; source images are not persisted.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [files]
              properties:
                files:
                  type: array
                  minItems: 1
                  maxItems: 20
                  items: { type: string, format: binary, contentMediaType: image/jpeg }
      responses: { "201": { $ref: "#/components/responses/ProcessingResult" }, default: { $ref: "#/components/responses/Error" } }
components:
  parameters:
    DocumentId:
      name: documentId
      in: path
      required: true
      schema: { type: string, format: uuid }
  requestBodies:
    BulkDocuments:
      required: true
      content:
        application/json:
          schema:
            type: object
            required: [documentIds]
            properties:
              documentIds:
                type: array
                minItems: 1
                maxItems: 200
                items: { type: string, format: uuid }
                description: Repeated identifiers are collapsed to a single removal.
...
    PageRange:
      required: true
      content:
        application/json:
          schema:
            allOf:
              - { $ref: "#/components/schemas/SingleDocumentOperation" }
              - type: object
                required: [ranges]
                properties:
                  ranges: { type: string, pattern: "^[0-9,-]+$", example: "1-3,5" }
    PageList:
      required: true
      content:
        application/json:
          schema:
            allOf:
              - { $ref: "#/components/schemas/SingleDocumentOperation" }
              - type: object
                required: [pages]
                properties: { pages: { type: array, minItems: 1, items: { type: integer, minimum: 1 } } }
  responses:
    Error:
      description: Stable error envelope
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
    ProcessingResult:
      description: Output committed as a new immutable version
      content:
        application/json:
          schema: { $ref: "#/components/schemas/DirectToolResult" }
    SplitProcessingResult:
      description: One immutable one-page PDF output for every selected source page
      content:
        application/json:
          schema:
            type: object
            required: [results, savedToRecent]
            properties:
              results:
                type: array
                minItems: 1
                items: { $ref: "#/components/schemas/DirectToolResult" }
              savedToRecent: { type: boolean }
    BulkDeleteResult:
      description: Per-document outcome for one bulk deletion
      content:
        application/json:
          schema: { $ref: "#/components/schemas/BulkDeleteResult" }
  schemas:
...
    RegisteredFont:
      type: object
      required: [id, family, serif, fixed]
      properties:
        id: { type: string, description: Identifier used by AnnotationText.font }
        family: { type: string, description: PostScript name of the face }
        serif: { type: boolean }
        fixed: { type: boolean }
...
    AnnotationDocument:
      type: object
      required: [pages]
      description: >-
        Coordinates are PDF user-space points with the origin at the bottom-left
        of the page. At most 500 objects per page and 5000 in total.
      properties:
        pages:
          type: array
          items: { $ref: "#/components/schemas/AnnotationPage" }
    DirectToolResult:
      type: object
      required: [document, version, downloadUrl, beforeBytes, afterBytes, savedToRecent]
      properties:
        document: { $ref: "#/components/schemas/Document" }
        version: { $ref: "#/components/schemas/DocumentVersion" }
        downloadUrl: { type: string, example: "/api/documents/00000000-0000-0000-0000-000000000000/versions/00000000-0000-0000-0000-000000000000/content" }
        beforeBytes: { type: integer, format: int64 }
        afterBytes: { type: integer, format: int64 }
        savedToRecent: { type: boolean, description: The current local persistence implementation records direct uploads internally. }
        outputName: { type: string, description: Suggested download filename for an individual output. }
...
    EditSessionEnvelope:
      type: object
      required: [session, document]
      properties:
        document: { $ref: "#/components/schemas/Document" }
        session:
          type: object
          required: [id, filename, byteSize, mode, previewUrl, downloadUrl]
          properties:
            id: { type: string, format: uuid }
            filename: { type: string }
            byteSize: { type: integer, format: int64 }
            pageCount: { type: [integer, "null"] }
            mode: { type: string, const: preview }
            previewUrl: { type: string }
            downloadUrl: { type: string }
    Document:
      type: object
      required: [id, originalName, mediaType, byteSize, checksumSha256, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        originalName: { type: string }
        mediaType: { type: string, const: application/pdf }
        byteSize: { type: integer, format: int64 }
        pageCount: { type: [integer, "null"], minimum: 1 }
        checksumSha256: { type: string, pattern: "^[a-f0-9]{64}$" }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        deletedAt: { type: [string, "null"], format: date-time }
    DocumentVersion:
      type: object
      required: [id, documentId, parentVersionId, operation, byteSize, checksumSha256, metadata, createdAt]
      properties:
        id: { type: string, format: uuid }
        documentId: { type: string, format: uuid }
        parentVersionId: { type: [string, "null"], format: uuid }
        operation: { type: string }
        byteSize: { type: integer, format: int64 }
        checksumSha256: { type: string }
        metadata: { type: object, additionalProperties: true }
        createdAt: { type: string, format: date-time }
    SingleDocumentOperation:
      type: object
      required: [documentId]
      properties:
        documentId: { type: string, format: uuid }
        confirmSignatures: { type: boolean, description: Explicit acknowledgment that modification may invalidate digital signatures. }
    PDFMetadata:
      type: object
      required: [title, author, subject, keywords]
      properties:
        title: { type: string }
        author: { type: string }
        subject: { type: string }
        keywords: { type: string }
...